@@ -1,9 +1,9 @@
 using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Hosting;
-using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.SpaServices.AngularCli;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
+using Microsoft.Extensions.Hosting;
 using Sketch7.DotnetNgx.Core;
 
 namespace Sketch7.DotnetNgx
@@ -20,7 +20,7 @@ namespace Sketch7.DotnetNgx
 		// This method gets called by the runtime. Use this method to add services to the container.
 		public void ConfigureServices(IServiceCollection services)
 		{
-			services.AddMvc().SetCompatibilityVersion(CompatibilityVersion.Version_2_2);
+			services.AddControllersWithViews();
 			// services.AddNodeServices(options =>
 			// {
 			// 	options.LaunchWithDebugging = true;
@@ -34,7 +34,7 @@ namespace Sketch7.DotnetNgx
 		}
 
 		// This method gets called by the runtime. Use this method to configure the HTTP request pipeline.
-		public void Configure(IApplicationBuilder app, IHostingEnvironment env)
+		public void Configure(IApplicationBuilder app, IWebHostEnvironment env)
 		{
 			if (env.IsDevelopment())
 			{
@@ -49,13 +49,18 @@ namespace Sketch7.DotnetNgx
 
 			app.UseHttpsRedirection();
 			app.UseStaticFiles();
-			app.UseSpaStaticFiles();
-
-			app.UseMvc(routes =>
+			if (!env.IsDevelopment())
 			{
-				routes.MapRoute(
+				app.UseSpaStaticFiles();
+			}
+
+			app.UseRouting();
+
+			app.UseEndpoints(endpoints =>
+			{
+				endpoints.MapControllerRoute(
 					name: "default",
-					template: "{controller}/{action=Index}/{id?}");
+					pattern: "{controller}/{action=Index}/{id?}");
 			});
 
 			app.UseSpa(spa =>
